pyproject.toml
[project]
name = "quiz-master-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.8.1",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
